Handle exceptions from initial fiber gracefully
Currently, if an exception is thrown in the initial fiber, RUN-FIBERS never
returns, because the atomic box RET is never written to. This change makes it so
RET is written to in case of both normal and exceptional exits, and any caught
exception is re-thrown in the context of the caller.